modelling/requirements.txt
mlflow==2.18.0
tensorflow==2.13.0
pandas
numpy
scikit-learn
nltk
Sastrawi
requests
matplotlib
seaborn
rapidfuzz
imbalanced-learn
pyyaml